Move Route package, Protocol, RequestMethod & StatusCode under httpserver
