+++ src/main/java/httpserver/RequestMethod.java
@@ -1,4 +1,4 @@
-package httpserver.http;
+package httpserver;
 
 public enum RequestMethod {
     GET,
