+++ src/main/java/httpserver/route/requestmethod/URI.java
@@ -1,4 +1,4 @@
-package httpserver.http.route.requestmethod;
+package httpserver.route.requestmethod;
 
 import httpserver.http.request.Request;
