+++ src/test/java/httpserver/http/response/ResponseTest.java
@@ -1,7 +1,7 @@
 package httpserver.http.response;
 
-import httpserver.http.Protocol;
-import httpserver.http.StatusCode;
+import httpserver.Protocol;
+import httpserver.StatusCode;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
