+++ src/test/java/httpserver/route/requestmethod/URITest.java
@@ -1,4 +1,4 @@
-package httpserver.http.route.requestmethod;
+package httpserver.route.requestmethod;
 
 import httpserver.http.request.RequestFactory;
 import org.junit.jupiter.api.Test;
